Use postgres as the Target for both Redshift and Postgres (only add an attribute use_db_timestamps). And rename postgreslike to rdbms
